finished email selection page
