+++ imports/ui/Summary/Table/DetailsList.jsx
@@ -1,0 +1,57 @@
+import React, { Component } from 'react';
+
+export default class DetailsList extends Component {
+  constructor(props) {
+    super(props);
+    this.state = ({
+      show: false
+    });
+  }
+
+  showClicked() {
+    this.setState({
+      show: true
+    });
+  }
+
+  hideClicked() {
+    this.setState({
+      show: false
+    });
+  }
+
+  render() {
+    var questionComponents = [];
+    for (var i = 0; i < this.props.questions.length; i++) {
+      questionComponents.push(
+        <div key={i}>
+          <p>{this.props.questions[i].question}</p>
+        </div>
+      );
+    }
+
+    if (this.state.show) {
+      return (
+        <div className="question-summary">
+          <div>
+            <h4>Questions</h4>
+            {questionComponents}
+          </div>
+          <div className="showhide-details" onClick={() => this.hideClicked()}>
+            Close Details
+          </div>
+        </div>
+      );
+    }
+    else {
+      return (
+        <div className="question-summary">
+          <div className="showhide-details" onClick={() => this.showClicked()}>
+            Show Details
+          </div>
+        </div>
+      );
+    }
+
+  }
+}
